sms-agent/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pyautogui==0.9.54
pillow==10.1.0